feat: adiciona validação para remoção de categoria e tratamento de exceção

+++ Domain/Commands/Categoria/RemoverCategoriaCommand.cs
@@ -1,5 +1,6 @@
 ﻿using Crosscutting.Constantes;
 using Crosscutting.Exceptions;
+using Domain.Interfaces;
 using Domain.Repositories;
 using MediatR;
 
@@ -10,7 +11,7 @@ public class RemoverCategoriaCommand : IRequest
     public Guid CategoriaId { get; set; }
 }
 
-public class RemoverCategoriaCommandHandler(ICategoriaRepository repository) : IRequestHandler<RemoverCategoriaCommand>
+public class RemoverCategoriaCommandHandler(ICategoriaRepository repository, ICategoriaService service) : IRequestHandler<RemoverCategoriaCommand>
 {
     public async Task Handle(RemoverCategoriaCommand request, CancellationToken cancellationToken)
     {
@@ -19,6 +20,9 @@ public class RemoverCategoriaCommandHandler(ICategoriaRepository repository) : I
         
         var categoria = await repository.ObterPorIdAsync(request.CategoriaId) 
             ?? throw new NaoEncontradoException(ErrorMessages.NaoExiste(Entidades.Categoria));
+        
+        if (!await service.PodeRemoverCategoria(request.CategoriaId))
+            throw new RegraDeNegocioException(ErrorMessages.CategoriaPossuiProdutos);
         
         await repository.RemoverESalvarAsync(categoria);
     }
